Debugged regen, contact, et script destroyable
Tous les porteurs du script destroyable prennent maintenant des dommages, et une regeneration de santé correcte.
Le script Destroyable est maintenant plus souple sur son recuperage du script canDamage ce qui permet de minimiser les bugs

+++ Assets/Scripts/ScriptableObjects/ObjectsData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class ObjectsData : ScriptableObject
 {
     [Header("Basic Stats")]
-    public int maxHealth;
-    public int regenAmount;
+    public float maxHealth;
+    public float regenAmount;
     public float regenSpeed;
     public float regenFirstWait;
 
@@ -15,13 +15,13 @@ public class ObjectsData : ScriptableObject
     public int goldLoot;
 
     [Header("Burning Stats")]
-    public int burningAmount;
+    public float burningAmount;
     public float burningSpeed;
 
     [Header("Character Stats")]
-    public int walkSpeed;
+    public float walkSpeed;
     public float outOfRangeDetachTime;
-    public int damageAmount;
+    public float damageAmount;
     public float damageCooldown;
     public float stunDuration;
